Added Enemy
Added Enemy, meta-inf files, methods for color on entities and added enemy processes to the Game class
