--- AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/events/Event.java
@@ -1,20 +1,0 @@
-package dk.sdu.mmmi.cbse.common.events;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import java.io.Serializable;
-
-/**
- *
- * @author Mads
- */
-public class Event implements Serializable{
-    private final Entity source;
-
-    public Event(Entity source) {
-        this.source = source;
-    }
-
-    public Entity getSource() {
-        return source;
-    }
-}
