--- AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/services/IEntityProcessingService.java
@@ -1,9 +1,0 @@
-package dk.sdu.mmmi.cbse.common.services;
-
-import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.World;
-
-public interface IEntityProcessingService {
-
-    void process(GameData gameData, World world);
-}
