--- AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/services/IGamePluginService.java
@@ -1,10 +1,0 @@
-package dk.sdu.mmmi.cbse.common.services;
-
-import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.World;
-
-public interface IGamePluginService {
-    void start(GameData gameData, World world);
-
-    void stop(GameData gameData, World world);
-}
