--- AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/cbse/common/services/IPostEntityProcessingService.java
@@ -1,12 +1,0 @@
-package dk.sdu.mmmi.cbse.common.services;
-
-import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.World;
-
-/**
- *
- * @author jcs
- */
-public interface IPostEntityProcessingService  {
-        void process(GameData gameData, World world);
-}
