+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/perib21/common/data/entityparts/MovingPart.java
@@ -1,0 +1,119 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package dk.sdu.mmmi.perib21.common.data.entityparts;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+import dk.sdu.mmmi.perib21.common.data.GameData;
+
+import static java.lang.Math.cos;
+import static java.lang.Math.sin;
+import static java.lang.Math.sqrt;
+
+/**
+ *
+ * @author Alexander
+ */
+public class MovingPart
+        implements EntityPart {
+
+    private float dx, dy;
+    private float deceleration, acceleration;
+    private float maxSpeed, rotationSpeed;
+    private boolean left, right, up;
+
+    public MovingPart(float deceleration, float acceleration, float maxSpeed, float rotationSpeed) {
+        this.deceleration = deceleration;
+        this.acceleration = acceleration;
+        this.maxSpeed = maxSpeed;
+        this.rotationSpeed = rotationSpeed;
+    }
+
+    public void setDeceleration(float deceleration) {
+        this.deceleration = deceleration;
+    }
+
+    public void setAcceleration(float acceleration) {
+        this.acceleration = acceleration;
+    }
+
+    public void setMaxSpeed(float maxSpeed) {
+        this.maxSpeed = maxSpeed;
+    }
+
+    public void setRotationSpeed(float rotationSpeed) {
+        this.rotationSpeed = rotationSpeed;
+    }
+
+    public void setLeft(boolean left) {
+        this.left = left;
+    }
+
+    public void setRight(boolean right) {
+        this.right = right;
+    }
+
+    public void setUp(boolean up) {
+        this.up = up;
+    }
+
+    @Override
+    public void process(GameData gameData, Entity entity) {
+        PositionPart positionPart = entity.getPart(PositionPart.class);
+        float x = positionPart.getX();
+        float y = positionPart.getY();
+        float radians = positionPart.getRadians();
+        float dt = gameData.getDelta();
+
+        // turning
+        if (left) {
+            radians += rotationSpeed * dt;
+        }
+
+        if (right) {
+            radians -= rotationSpeed * dt;
+        }
+
+        // accelerating            
+        if (up) {
+            dx += cos(radians) * acceleration * dt;
+            dy += sin(radians) * acceleration * dt;
+        }
+
+        // deccelerating
+        float vec = (float) sqrt(dx * dx + dy * dy);
+        if (vec > 0) {
+            dx -= (dx / vec) * deceleration * dt;
+            dy -= (dy / vec) * deceleration * dt;
+        }
+        if (vec > maxSpeed) {
+            dx = (dx / vec) * maxSpeed;
+            dy = (dy / vec) * maxSpeed;
+        }
+
+        // set position
+        x += dx * dt;
+        if (x > gameData.getDisplayWidth()) {
+            x = 0;
+        }
+        else if (x < 0) {
+            x = gameData.getDisplayWidth();
+        }
+
+        y += dy * dt;
+        if (y > gameData.getDisplayHeight()) {
+            y = 0;
+        }
+        else if (y < 0) {
+            y = gameData.getDisplayHeight();
+        }
+
+        positionPart.setX(x);
+        positionPart.setY(y);
+
+        positionPart.setRadians(radians);
+    }
+
+}
