+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/perib21/common/events/Event.java
@@ -1,0 +1,20 @@
+package dk.sdu.mmmi.perib21.common.events;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+import java.io.Serializable;
+
+/**
+ *
+ * @author Mads
+ */
+public class Event implements Serializable{
+    private final Entity source;
+
+    public Event(Entity source) {
+        this.source = source;
+    }
+
+    public Entity getSource() {
+        return source;
+    }
+}
