+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/perib21/common/services/IGamePluginService.java
@@ -1,0 +1,10 @@
+package dk.sdu.mmmi.perib21.common.services;
+
+import dk.sdu.mmmi.perib21.common.data.GameData;
+import dk.sdu.mmmi.perib21.common.data.World;
+
+public interface IGamePluginService {
+    void start(GameData gameData, World world);
+
+    void stop(GameData gameData, World world);
+}
