+++ AsteroidsEntityFramework/Common/src/main/java/dk/sdu/mmmi/perib21/common/services/IPostEntityProcessingService.java
@@ -1,0 +1,12 @@
+package dk.sdu.mmmi.perib21.common.services;
+
+import dk.sdu.mmmi.perib21.common.data.GameData;
+import dk.sdu.mmmi.perib21.common.data.World;
+
+/**
+ *
+ * @author jcs
+ */
+public interface IPostEntityProcessingService  {
+        void process(GameData gameData, World world);
+}
