--- AsteroidsEntityFramework/Core/src/main/java/dk/sdu/mmmi/cbse/managers/GameInputProcessor.java
@@ -1,80 +1,0 @@
-package dk.sdu.mmmi.cbse.managers;
-
-import com.badlogic.gdx.Input.Keys;
-import com.badlogic.gdx.InputAdapter;
-import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.GameKeys;
-
-public class GameInputProcessor extends InputAdapter {
-
-    private final GameData gameData;
-
-    public GameInputProcessor(GameData gameData) {
-        this.gameData = gameData;
-    }
-
-	public boolean keyDown(int k) {
-		if(k == Keys.UP) {
-                    gameData.getKeys().setKey(GameKeys.UP, true);
-		}
-		if(k == Keys.LEFT) {
-                    gameData.getKeys().setKey(GameKeys.LEFT, true);
-		}
-		if(k == Keys.DOWN) {
-                    gameData.getKeys().setKey(GameKeys.DOWN, true);
-		}
-		if(k == Keys.RIGHT) {
-                    gameData.getKeys().setKey(GameKeys.RIGHT, true);
-		}
-		if(k == Keys.ENTER) {
-                    gameData.getKeys().setKey(GameKeys.ENTER, true);
-		}
-		if(k == Keys.ESCAPE) {
-                    gameData.getKeys().setKey(GameKeys.ESCAPE, true);
-		}
-		if(k == Keys.SPACE) {
-                    gameData.getKeys().setKey(GameKeys.SPACE, true);
-		}
-		if(k == Keys.SHIFT_LEFT || k == Keys.SHIFT_RIGHT) {
-                    gameData.getKeys().setKey(GameKeys.SHIFT, true);
-		}
-		return true;
-	}
-	
-	public boolean keyUp(int k) {
-		if(k == Keys.UP) {
-                    gameData.getKeys().setKey(GameKeys.UP, false);
-		}
-		if(k == Keys.LEFT) {
-                    gameData.getKeys().setKey(GameKeys.LEFT, false);
-		}
-		if(k == Keys.DOWN) {
-                    gameData.getKeys().setKey(GameKeys.DOWN, false);
-		}
-		if(k == Keys.RIGHT) {
-                    gameData.getKeys().setKey(GameKeys.RIGHT, false);
-		}
-		if(k == Keys.ENTER) {
-                    gameData.getKeys().setKey(GameKeys.ENTER, false);
-		}
-		if(k == Keys.ESCAPE) {
-                    gameData.getKeys().setKey(GameKeys.ESCAPE, false);
-		}
-		if(k == Keys.SPACE) {
-                    gameData.getKeys().setKey(GameKeys.SPACE, false);
-		}
-		if(k == Keys.SHIFT_LEFT || k == Keys.SHIFT_RIGHT) {
-                    gameData.getKeys().setKey(GameKeys.SHIFT, false);
-		}
-		return true;
-	}
-	
-}
-
-
-
-
-
-
-
-
