+++ AsteroidsEntityFramework/Core/src/main/java/dk/sdu/mmmi/perib21/main/Main.java
@@ -1,0 +1,22 @@
+package dk.sdu.mmmi.perib21.main;
+
+import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
+import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+
+public class Main {
+	
+	public static void main(String[] args) {
+		
+		LwjglApplicationConfiguration cfg =
+			new LwjglApplicationConfiguration();
+		cfg.title = "Asteroids";
+		cfg.width = 500;
+		cfg.height = 400;
+		cfg.useGL30 = false;
+		cfg.resizable = false;
+		
+		new LwjglApplication(new Game(), cfg);
+		
+	}
+	
+}
