+++ AsteroidsEntityFramework/Enemy/src/main/java/dk/sdu/mmmi/perib21/enemysystem/Enemy.java
@@ -1,0 +1,6 @@
+package dk.sdu.mmmi.perib21.enemysystem;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+
+public class Enemy extends Entity {
+}
