+++ AsteroidsEntityFramework/Enemy/src/main/java/dk/sdu/mmmi/perib21/enemysystem/EnemyControlSystem.java
@@ -1,0 +1,61 @@
+package dk.sdu.mmmi.perib21.enemysystem;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+import dk.sdu.mmmi.perib21.common.data.GameData;
+import dk.sdu.mmmi.perib21.common.data.World;
+import dk.sdu.mmmi.perib21.common.data.entityparts.LifePart;
+import dk.sdu.mmmi.perib21.common.data.entityparts.MovingPart;
+import dk.sdu.mmmi.perib21.common.data.entityparts.PositionPart;
+import dk.sdu.mmmi.perib21.common.services.IEntityProcessingService;
+
+public class EnemyControlSystem implements IEntityProcessingService {
+    @Override
+    public void process(GameData gameData, World world) {
+
+        for (Entity enemy : world.getEntities(Enemy.class)) {
+
+            PositionPart positionPart = enemy.getPart(PositionPart.class);
+            MovingPart movingPart = enemy.getPart(MovingPart.class);
+            LifePart lifePart = enemy.getPart(LifePart.class);
+
+            movingPart.setLeft(false);
+            movingPart.setRight(true);
+            movingPart.setUp(true);
+
+
+            movingPart.process(gameData, enemy);
+            positionPart.process(gameData, enemy);
+            //lifePart.process(gameData,enemy);
+
+            updateShape(enemy);
+        }
+
+    }
+
+    private void updateShape(Entity enemy) {
+        float[] shapex = enemy.getShapeX();
+        float[] shapey = enemy.getShapeY();
+        PositionPart positionPart = enemy.getPart(PositionPart.class);
+        float x = positionPart.getX();
+        float y = positionPart.getY();
+        float radians = positionPart.getRadians();
+
+        shapex[0] = (float) (x + Math.cos(radians) * 8);
+        shapey[0] = (float) (y + Math.sin(radians) * 8);
+
+        shapex[1] = (float) (x + Math.cos(radians - 4 * 3.1415f / 5) * 8);
+        shapey[1] = (float) (y + Math.sin(radians - 4 * 3.1145f / 5) * 8);
+
+        shapex[2] = (float) (x + Math.cos(radians + 3.1415f) * 5);
+        shapey[2] = (float) (y + Math.sin(radians + 3.1415f) * 5);
+
+        shapex[3] = (float) (x + Math.cos(radians + 4 * 3.1415f / 5) * 8);
+        shapey[3] = (float) (y + Math.sin(radians + 4 * 3.1415f / 5) * 8);
+
+        enemy.setShapeX(shapex);
+        enemy.setShapeY(shapey);
+
+    }
+
+
+}
