+++ AsteroidsEntityFramework/Enemy/src/main/resources/META-INF/dk.sdu.mmmi.cbse.common.services.IEntityProcessingService
@@ -1,0 +1,1 @@
+dk.sdu.mmmi.perib21.enemySystem.EnemyControlSystem
