+++ AsteroidsEntityFramework/Enemy/src/main/resources/META-INF/dk.sdu.mmmi.cbse.common.services.IGamePluginService
@@ -1,0 +1,1 @@
+dk.sdu.mmmi.perib21.enemySystem.EnemyPlugin
