--- AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/cbse/playersystem/Player.java
@@ -1,12 +1,0 @@
-package dk.sdu.mmmi.cbse.playersystem;
-
-import dk.sdu.mmmi.cbse.common.data.Entity;
-import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
-
-/**
- *
- * @author corfixen
- */
-public class Player extends Entity {
-
-}
