+++ AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/perib21/playersystem/Player.java
@@ -1,0 +1,11 @@
+package dk.sdu.mmmi.perib21.playersystem;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+
+/**
+ *
+ * @author corfixen
+ */
+public class Player extends Entity {
+
+}
