+++ AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/perib21/playersystem/PlayerPlugin.java
@@ -1,0 +1,49 @@
+package dk.sdu.mmmi.perib21.playersystem;
+
+import dk.sdu.mmmi.perib21.common.data.Entity;
+import dk.sdu.mmmi.perib21.common.data.GameData;
+import dk.sdu.mmmi.perib21.common.data.World;
+import dk.sdu.mmmi.perib21.common.data.entityparts.MovingPart;
+import dk.sdu.mmmi.perib21.common.data.entityparts.PositionPart;
+import dk.sdu.mmmi.perib21.common.services.IGamePluginService;
+
+public class PlayerPlugin implements IGamePluginService {
+
+    private Entity player;
+
+    public PlayerPlugin() {
+    }
+
+    @Override
+    public void start(GameData gameData, World world) {
+        
+        // Add entities to the world
+        player = createPlayerShip(gameData);
+        world.addEntity(player);
+
+    }
+
+    private Entity createPlayerShip(GameData gameData) {
+
+        float deacceleration = 10;
+        float acceleration = 200;
+        float maxSpeed = 300;
+        float rotationSpeed = 5;
+        float x = gameData.getDisplayWidth() / 2;
+        float y = gameData.getDisplayHeight() / 2;
+        float radians = 3.1415f / 2;
+        
+        Entity playerShip = new Player();
+        playerShip.add(new MovingPart(deacceleration, acceleration, maxSpeed, rotationSpeed));
+        playerShip.add(new PositionPart(x, y, radians));
+
+        return playerShip;
+    }
+
+    @Override
+    public void stop(GameData gameData, World world) {
+        // Remove entities
+        world.removeEntity(player);
+    }
+
+}
