+++ AsteroidsEntityFramework/Player/src/main/resources/META-INF/services/dk.sdu.mmmi.perib21.common.services.IEntityProcessingService
@@ -1,0 +1,1 @@
+dk.sdu.mmmi.perib21.playersystem.PlayerControlSystem
