+++ AsteroidsEntityFramework/Player/src/main/resources/META-INF/services/dk.sdu.mmmi.perib21.common.services.IGamePluginService
@@ -1,0 +1,1 @@
+dk.sdu.mmmi.perib21.playersystem.PlayerPlugin
